# Conflicts:
#	ts/WoltLabSuite/Core/Notification/Handler.ts
#	wcfsetup/install/files/js/WoltLabSuite/Core/Notification/Handler.js
#	wcfsetup/install/files/lib/system/api/composer.json
#	wcfsetup/install/files/lib/system/api/composer.lock
#	wcfsetup/install/files/lib/system/api/composer/installed.json
#	wcfsetup/install/files/lib/system/api/composer/installed.php

+++ wcfsetup/install/files/lib/data/style/StyleAction.class.php
@@ -14,6 +14,7 @@ use wcf\system\image\ImageHandler;
 use wcf\system\language\LanguageFactory;
 use wcf\system\Regex;
 use wcf\system\request\LinkHandler;
+use wcf\system\style\command\CreateManifest;
 use wcf\system\style\StyleHandler;
 use wcf\system\WCF;
 use wcf\util\FileUtil;
@@ -324,7 +325,6 @@ class StyleAction extends AbstractDatabaseObjectAction implements IToggleAction
             'mstile-150x150.png' => 150,
         ];
 
-        $hasFavicon = $style->hasFavicon;
         if (\array_key_exists('favicon', $this->parameters['uploads'])) {
             /** @var \wcf\system\file\upload\UploadFile $file */
             $file = $this->parameters['uploads']['favicon'];
@@ -359,8 +359,6 @@ class StyleAction extends AbstractDatabaseObjectAction implements IToggleAction
                     (new StyleEditor($style))->update([
                         'hasFavicon' => 1,
                     ]);
-
-                    $hasFavicon = true;
                 }
             } else {
                 foreach ($images as $filename => $length) {
@@ -371,8 +369,14 @@ class StyleAction extends AbstractDatabaseObjectAction implements IToggleAction
                 if (\file_exists($style->getAssetPath() . "favicon.ico")) {
                     \unlink($style->getAssetPath() . "favicon.ico");
                 }
+                foreach (\glob($style->getAssetPath() . "manifest-*.json") as $filename) {
+                    \unlink($filename);
+                }
+                // delete the manifest.json generated before WSC 6.1
+                if (\file_exists($style->getAssetPath() . "manifest.json")) {
+                    \unlink($style->getAssetPath() . "manifest.json");
+                }
 
-                \unlink($style->getAssetPath() . "manifest.json");
                 \unlink($style->getAssetPath() . "browserconfig.xml");
                 foreach (['png', 'jpg', 'gif'] as $extension) {
                     if (\file_exists($style->getAssetPath() . "favicon-template." . $extension)) {
@@ -382,40 +386,14 @@ class StyleAction extends AbstractDatabaseObjectAction implements IToggleAction
                 (new StyleEditor($style))->update([
                     'hasFavicon' => 0,
                 ]);
-
-                $hasFavicon = false;
             }
         }
+        // need to reload the style object to get the updated hasFavicon value
+        $style = new Style($style->styleID);
+        $command = new CreateManifest($style);
+        $command();
 
-        if ($hasFavicon) {
-            // update manifest.json
-            $manifest = <<<'MANIFEST'
-{
-    "name": "",
-    "icons": [
-        {
-            "src": "android-chrome-192x192.png",
-            "sizes": "192x192",
-            "type": "image/png"
-        },
-        {
-            "src": "android-chrome-256x256.png",
-            "sizes": "256x256",
-            "type": "image/png"
-        },
-        {
-            "src": "android-chrome-512x512.png",
-            "sizes": "512x512",
-            "type": "image/png"
-        }
-    ],
-    "theme_color": "#ffffff",
-    "background_color": "#ffffff",
-    "display": "standalone"
-}
-MANIFEST;
-            \file_put_contents($style->getAssetPath() . "manifest.json", $manifest);
-
+        if ($style->hasFavicon) {
             $style->loadVariables();
             $tileColor = $style->getVariable('wcfHeaderBackground', true);
